Move to src dir layout
