+++ setup.cfg
@@ -39,13 +39,12 @@ ignore =
     # W503:  line break before binary operator (this is no longer PEP8 compliant)
     W503,
 per-file-ignores =
-    # Top level init file improves user experience by short cutting imports,
-    # this is not used in the package online by calling clients
+    # Package level init files improve user experience by short cutting imports.
     # F401: imported but unused
-    mbed_tools/__init__.py:F401
+    __init__.py:F401
     # Don't require docstrings in tests.
     # We evaluate the need for them on case by case basis.
-    test_*.py:D1
+    tests/*.py:D1
 max-line-length = 120
 docstring-convention = google
 copyright-check = True
