+++ src/mbed_tools/cli/__init__.py
@@ -1,0 +1,7 @@
+#
+# Copyright (C) 2020 Arm Mbed. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+"""Exposes the mbed_tools command line interface."""
+
+from mbed_tools.cli.cli import cli, LOGGER
