+++ src/mbed_tools/cli/_internal/__init__.py
@@ -1,0 +1,5 @@
+#
+# Copyright (C) 2020 Arm Mbed. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+"""mbed_tools cli code not to be accessed by external applications."""
