+++ src/mbed_tools/cli/cli.py
@@ -9,11 +9,11 @@ from pkg_resources import working_set
 from typing import Union, Any
 import click
 
-from mbed_tools_lib.logging import set_log_level, MbedToolsHandler
+from mbed_tools.lib.logging import set_log_level, MbedToolsHandler
 
 from mbed_build.mbed_tools import configure
 from mbed_devices.mbed_tools import cli as mbed_devices_cli
-from mbed_tools._internal.env_cli import cli as env_cli
+from mbed_tools.cli._internal.env_cli import cli as env_cli
 from mbed_project.mbed_tools.cli import init, clone, checkout, libs
