+++ tests/__init__.py
@@ -2,4 +2,3 @@
 # Copyright (C) 2020 Arm Mbed. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
-"""Code not to be accessed by external applications."""
